Had to re-add the RAG_system_project
After pushing the RAG_system_project to Github as a submodule and not realizing it would not display properly online, I decided to remove the submodule. Using the 'git submodule deinit' command accidently deleted the entire project from the working directory. I had to restore the project from the local repo it was originally created in.